requirements.txt
# Core
beautifulsoup4>=4.12.0
lxml>=4.9.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.0.0
rich>=13.0.0

# Búsqueda y Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Interfaz Chat IA
gradio>=4.0.0

# LLM via OpenRouter
httpx>=0.27.0

# Telethon (para sincronización futura)
telethon>=1.28.0
